# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: History Wiki API
  version: 0.1.0
  description: 历史 Wiki 的公开读者 API。空数据库行为是本契约的一等状态。
  license:
    name: Unlicensed
servers:
  - url: /
security: []
paths:
...
  /api/v1/admin/regions:
    get:
      operationId: listRegions
      summary: 列出或搜索地区
      description: 编辑者和管理员均可查询地区；名称相同的地区通过消歧名称和稳定 ID 区分。
      security:
        - sessionCookie: []
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
          description: 在当前名称和消歧名称中执行不区分拉丁文字大小写的子串搜索。
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/CanonicalEntityStatus" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: 地区列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegionListResponse" }
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createRegion
      summary: 创建地区
      description: 编辑者和管理员均可创建地区；同一账号重复使用相同幂等键和请求内容时返回首次创建的地区。
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRegionRequest" }
      responses:
        "201":
          description: 地区已创建；幂等重放仍返回相同状态和实体，并设置 Idempotency-Replayed。
          headers:
            ETag:
              schema: { type: string }
            Location:
              schema: { type: string }
            Idempotency-Replayed:
              schema: { type: string, enum: ["true"] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Region" }
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/admin/regions/{regionId}:
    patch:
      operationId: updateRegion
      summary: 修改地区
      description: 只有管理员可以修改当前名称、消歧名称或有效状态；使用强 ETag 防止并发静默覆盖。
      security:
        - sessionCookie: []
      parameters:
        - name: regionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRegionRequest" }
      responses:
        "200":
          description: 地区已修改
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Region" }
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "428":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-history_wiki_session
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: 必须与 Secure CSRF Cookie 一致，并与当前服务端 Session 绑定。
      schema: { type: string, minLength: 1 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 同一账号和写操作范围内唯一；重试时必须复用相同值与相同请求内容。
      schema: { type: string, minLength: 1, maxLength: 200 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 使用最近一次资源响应中的强 ETag，例如 `"1"`。
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PayloadTooLarge:
      description: JSON 请求体超过此操作允许的 2 MiB 或 20 MiB 上限
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: 请求超过约定速率
      headers:
        Retry-After:
          description: 建议重试前等待的秒数
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    CanonicalEntityStatus:
      type: string
      enum: [active, inactive, merged]
    Region:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - disambiguationLabel
        - status
        - mergedIntoId
        - lockVersion
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        disambiguationLabel:
          type: [string, "null"]
          maxLength: 120
        status: { $ref: "#/components/schemas/CanonicalEntityStatus" }
        mergedIntoId: { type: [string, "null"], format: uuid }
        lockVersion: { type: integer, format: int64, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RegionListResponse:
      type: object
      additionalProperties: false
      required: [regions]
      properties:
        regions:
          type: array
          items: { $ref: "#/components/schemas/Region" }
    CreateRegionRequest:
      type: object
      additionalProperties: false
      required: [name, disambiguationLabel]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        disambiguationLabel:
          type: [string, "null"]
          maxLength: 120
    UpdateRegionRequest:
      type: object
      additionalProperties: false
      required: [name, disambiguationLabel, status]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        disambiguationLabel:
          type: [string, "null"]
          maxLength: 120
        status:
          type: string
          enum: [active, inactive]
...
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        requestId: { type: string }
...